\name{FunWithNumbers-package}
\alias{FunWithNumbers-package}
\alias{FunWithNumbers}
\docType{package}
\title{
\packageTitle{FunWithNumbers}
}
\description{
\packageDescription{FunWithNumbers}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{FunWithNumbers}
\packageIndices{FunWithNumbers}

}
\author{
\packageAuthor{FunWithNumbers}

Maintainer: \packageMaintainer{FunWithNumbers}
}

\keyword{ package }

