\name{preciseNumbersAsChar}
\alias{charE}
\alias{charPhi}
\alias{charPi}
\alias{charRoot2}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
 High-precision values for some common constants, in character strings. 
}
\description{
These are provided for use when playing around with some of the functions in this package, e.g., \code{bestFrac} or \code{cfrac} 
}

\details{
These represent, in order, "e" (natural log base), the golden ratio (1+sqrt(5))/2 aka "phi", "pi", and the square root of 2 as generated via \code{rmpfr} with 10 000 binary bits of precision.  There are many websites which can provide upwards of a million decimal digits for these constants for those who are interested. 
}


\author{
Carl Witthoft,  \email{carl@witthoft.com} 
}
